Params settings, Delay Buffer indexing and JUCE testing
Simplechorus has had some set() methods added to change the internal parameters.

Overloading of constructors has been added for JUCE compatiblity. Since most classes just need sample rate, this will need to be set beyond the constructor stage.

Delay indexing was broken and still is to a degree, there are a few problems with regards to modulating the delay buffer index and tieing this in with buffer size. CAUTION, unexpected terminations are likely rooted in this problem